refactor: Navbar
